move loaders to a single package, since they're not really part of any one pipeline
