+++ mowgli_etl/_loader.py
@@ -1,5 +1,8 @@
 import logging
+import subprocess
+import sys
 from abc import ABC, abstractmethod
+from pathlib import Path
 
 from mowgli_etl.model.edge import Edge
 from mowgli_etl.model.node import Node
@@ -9,6 +12,11 @@ from mowgli_etl.pipeline_storage import PipelineStorage
 class _Loader(ABC):
     def __init__(self):
         self._logger = logging.getLogger(self.__class__.__name__)
+
+    def _bzip_file(self, file_path: Path):
+        if sys.platform.startswith("win"):
+            return
+        subprocess.call(["bzip2", "-9", "-f", str(file_path)])
 
     @abstractmethod
     def close(self) -> None:
