+++ mowgli_etl/_path_loader.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+
+from mowgli_etl._loader import _Loader
+from mowgli_etl.model.path import Path
+
+
+class _PathLoader(_Loader):
+    @abstractmethod
+    def load_path(self, path: Path):
+        raise NotImplementedError
