+++ mowgli_etl/loader/cskg_csv/cskg_csv_loader.py
@@ -1,11 +1,11 @@
 from csv import DictWriter
+from pathlib import Path
 from typing import Dict, Callable
 
 from mowgli_etl._edge_loader import _EdgeLoader
 from mowgli_etl._node_loader import _NodeLoader
 from mowgli_etl.model.edge import Edge
 from mowgli_etl.model.node import Node
-from mowgli_etl._loader import _Loader
 
 
 class CskgCsvLoader(_EdgeLoader, _NodeLoader):
@@ -18,6 +18,10 @@ class CskgCsvLoader(_EdgeLoader, _NodeLoader):
         'aliases': lambda node: ' '.join(node.aliases) if node.aliases is not None else None,
         'other': lambda obj: str(obj.other) if obj.other is not None else None
     }
+
+    def __init__(self, *, bzip: bool = False):
+        _EdgeLoader.__init__(self)
+        self.__bzip = bzip
 
     def open(self, storage):
         self.__storage = storage
@@ -38,6 +42,9 @@ class CskgCsvLoader(_EdgeLoader, _NodeLoader):
     def close(self):
         self.__edge_file.close()
         self.__node_file.close()
+        if self.__bzip:
+            self._bzip_file(Path(self.__edge_file.name))
+            self._bzip_file(Path(self.__node_file.name))
 
     def load_edge(self, edge: Edge):
         self._write_csv_line(self.__edge_writer, self.__EDGE_CSV_FIELDS, edge)
