+++ tests/mowgli_etl_test/pipeline/rdf/quad_rdf_loader_test.py
@@ -2,7 +2,7 @@ import pytest
 import rdflib
 
 from mowgli_etl.model.edge import Edge
-from mowgli_etl.pipeline.rdf.quad_rdf_loader import QuadRdfLoader
+from mowgli_etl.loader.rdf.quad_rdf_loader import QuadRdfLoader
 from mowgli_etl.pipeline_storage import PipelineStorage
 from mowgli_etl.pipeline_wrapper import PipelineWrapper
 from tests.mowgli_etl_test.pipeline.rdf.rdf_loader_test_pipeline import RdfLoaderTestPipeline
